--- api/app/tests/unittests/test_models_vuln_matching_ecosystem.py
@@ -1,34 +1,0 @@
-import pytest
-
-from app import models
-
-
-class TestVulnMatchingEcosystem:
-    @pytest.mark.parametrize(
-        "ecosystem, package_type, expected",
-        [
-            ("alpine-3.22.0", "OS", "alpine-3.22"),
-            ("alpine-3", "OS", "alpine-3"),
-            ("alpine-test-3.22.0", "OS", "alpine-test-3.22.0"),
-            ("alpine", "OS", "alpine"),
-            ("ubuntu-20.04", "OS", "ubuntu-20.04"),
-            ("npm", "LANG", "npm"),
-            ("pypi", "LANG", "pypi"),
-        ],
-    )
-    def test_it_should_return_expected_value_for_vuln_matching_ecosystem(
-        self, ecosystem: str, package_type: str, expected: str
-    ):
-        # Given
-        if package_type == "OS":
-            os_package = models.OSPackage(
-                name="test_package", ecosystem=ecosystem, source_name="test_source_name_package"
-            )
-        else:
-            lang_package = models.LangPackage(name="test_package", ecosystem=ecosystem)
-
-        # Then
-        if package_type == "OS":
-            assert os_package.vuln_matching_ecosystem == expected
-        else:
-            assert lang_package.vuln_matching_ecosystem == expected
